[Refactor] 존재 여부 키 커스텀훅 리팩토링
