+++ src/features/accountItem/index.ts
@@ -1,4 +1,9 @@
 export type { AccountItem, MoneyType } from './model/types'
-export { fetchByMonth } from './service/accountItem'
+export {
+  fetchByMonth,
+  fetchAllItems,
+  createRecurringItem
+} from './service/accountItem'
 export { useInstallmentItem } from './model/useInstallmentItem'
 export { useRecurringItem } from './model/useRecurringItem'
+export { useExistingKey } from './model/useExistingKey'
